+++ src/index.ts
@@ -1,6 +1,7 @@
 export { doctorConfig, formatDoctor } from "./doctor";
 export { findConfigPath, loadConfig } from "./config";
 export { formatDecision } from "./format";
+export { starterConfig, writeStarterConfig } from "./init";
 export { getRunnerOption, providerFactories } from "./providers";
 export { displayUnit, quotaFor, roundQuota } from "./quota";
 export { resolveFreelane } from "./resolve";